refine tests and testkit test for scheduler
