+++ pkg/disttask/framework/BUILD.bazel
@@ -16,18 +16,13 @@ go_test(
     shard_count = 32,
     deps = [
         "//pkg/disttask/framework/dispatcher",
-        "//pkg/disttask/framework/dispatcher/mock",
         "//pkg/disttask/framework/handle",
-        "//pkg/disttask/framework/mock",
-        "//pkg/disttask/framework/mock/execute",
         "//pkg/disttask/framework/proto",
         "//pkg/disttask/framework/scheduler",
         "//pkg/disttask/framework/storage",
-        "//pkg/domain/infosync",
+        "//pkg/disttask/framework/testutil",
         "//pkg/testkit",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
-        "@com_github_tikv_client_go_v2//util",
-        "@org_uber_go_mock//gomock",
     ],
 )
