+++ pkg/disttask/framework/framework_err_handling_test.go
@@ -15,144 +15,35 @@
 package framework_test
 
 import (
-	"context"
-	"errors"
-	"sync"
 	"testing"
 
-	"github.com/pingcap/tidb/pkg/disttask/framework/dispatcher"
-	mockDispatch "github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/mock"
 	"github.com/pingcap/tidb/pkg/disttask/framework/proto"
-	"github.com/pingcap/tidb/pkg/domain/infosync"
-	"github.com/pingcap/tidb/pkg/testkit"
-	"github.com/tikv/client-go/v2/util"
-	"go.uber.org/mock/gomock"
+	"github.com/pingcap/tidb/pkg/disttask/framework/testutil"
 )
-
-var (
-	callTime = 0
-)
-
-func getPlanNotRetryableErrDispatcherExt(ctrl *gomock.Controller) dispatcher.Extension {
-	mockDispatcher := mockDispatch.NewMockExtension(ctrl)
-	mockDispatcher.EXPECT().OnTick(gomock.Any(), gomock.Any()).Return().AnyTimes()
-	mockDispatcher.EXPECT().GetEligibleInstances(gomock.Any(), gomock.Any()).DoAndReturn(
-		func(_ context.Context, _ *proto.Task) ([]*infosync.ServerInfo, bool, error) {
-			return generateSchedulerNodes4Test()
-		},
-	).AnyTimes()
-	mockDispatcher.EXPECT().IsRetryableErr(gomock.Any()).Return(false).AnyTimes()
-	mockDispatcher.EXPECT().GetNextStep(gomock.Any()).DoAndReturn(
-		func(task *proto.Task) proto.Step {
-			return proto.StepDone
-		},
-	).AnyTimes()
-	mockDispatcher.EXPECT().OnNextSubtasksBatch(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(
-		func(_ context.Context, _ dispatcher.TaskHandle, gTask *proto.Task, _ []*infosync.ServerInfo, _ proto.Step) (metas [][]byte, err error) {
-			return nil, errors.New("not retryable err")
-		},
-	).AnyTimes()
-
-	mockDispatcher.EXPECT().OnErrStage(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(
-		func(_ context.Context, _ dispatcher.TaskHandle, _ *proto.Task, _ []error) (meta []byte, err error) {
-			return nil, errors.New("not retryable err")
-		},
-	).AnyTimes()
-	return mockDispatcher
-}
-
-func getPlanErrDispatcherExt(ctrl *gomock.Controller) dispatcher.Extension {
-	mockDispatcher := mockDispatch.NewMockExtension(ctrl)
-	mockDispatcher.EXPECT().OnTick(gomock.Any(), gomock.Any()).Return().AnyTimes()
-	mockDispatcher.EXPECT().GetEligibleInstances(gomock.Any(), gomock.Any()).DoAndReturn(
-		func(_ context.Context, _ *proto.Task) ([]*infosync.ServerInfo, bool, error) {
-			return generateSchedulerNodes4Test()
-		},
-	).AnyTimes()
-	mockDispatcher.EXPECT().IsRetryableErr(gomock.Any()).Return(true).AnyTimes()
-	mockDispatcher.EXPECT().GetNextStep(gomock.Any()).DoAndReturn(
-		func(task *proto.Task) proto.Step {
-			switch task.Step {
-			case proto.StepInit:
-				return proto.StepOne
-			case proto.StepOne:
-				return proto.StepTwo
-			default:
-				return proto.StepDone
-			}
-		},
-	).AnyTimes()
-	mockDispatcher.EXPECT().OnNextSubtasksBatch(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(
-		func(_ context.Context, _ dispatcher.TaskHandle, gTask *proto.Task, _ []*infosync.ServerInfo, _ proto.Step) (metas [][]byte, err error) {
-			if gTask.Step == proto.StepInit {
-				if callTime == 0 {
-					callTime++
-					return nil, errors.New("retryable err")
-				}
-				return [][]byte{
-					[]byte("task1"),
-					[]byte("task2"),
-					[]byte("task3"),
-				}, nil
-			}
-			if gTask.Step == proto.StepOne {
-				return [][]byte{
-					[]byte("task4"),
-				}, nil
-			}
-			return nil, nil
-		},
-	).AnyTimes()
-
-	mockDispatcher.EXPECT().OnErrStage(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(
-		func(_ context.Context, _ dispatcher.TaskHandle, _ *proto.Task, _ []error) (meta []byte, err error) {
-			if callTime == 1 {
-				callTime++
-				return nil, errors.New("not retryable err")
-			}
-			return []byte("planErrTask"), nil
-		},
-	).AnyTimes()
-	return mockDispatcher
-}
 
 func TestPlanErr(t *testing.T) {
-	m := sync.Map{}
-	ctrl := gomock.NewController(t)
+	ctx, ctrl, test_context, distContext := testutil.InitTestContext(t, 2)
 	defer ctrl.Finish()
-	ctx := context.Background()
-	ctx = util.WithInternalSourceType(ctx, "dispatcher")
-
-	RegisterTaskMeta(t, ctrl, &m, getPlanErrDispatcherExt(ctrl))
-	distContext := testkit.NewDistExecutionContext(t, 2)
-	DispatchTaskAndCheckSuccess(ctx, "key1", t, &m)
-	callTime = 0
+	testutil.RegisterTaskMeta(t, ctrl, testutil.GetPlanErrDispatcherExt(ctrl, test_context), test_context)
+	testutil.DispatchTaskAndCheckSuccess(t, ctx, "key1", test_context)
+	test_context.CallTime = 0
 	distContext.Close()
 }
 
 func TestRevertPlanErr(t *testing.T) {
-	m := sync.Map{}
-
-	ctrl := gomock.NewController(t)
+	ctx, ctrl, test_context, distContext := testutil.InitTestContext(t, 2)
 	defer ctrl.Finish()
-	ctx := context.Background()
-	ctx = util.WithInternalSourceType(ctx, "dispatcher")
-
-	RegisterTaskMeta(t, ctrl, &m, getPlanErrDispatcherExt(ctrl))
-	distContext := testkit.NewDistExecutionContext(t, 2)
-	DispatchTaskAndCheckSuccess(ctx, "key1", t, &m)
+	testutil.RegisterTaskMeta(t, ctrl, testutil.GetPlanErrDispatcherExt(ctrl, test_context), test_context)
+	testutil.DispatchTaskAndCheckSuccess(t, ctx, "key1", test_context)
+	test_context.CallTime = 0
 	distContext.Close()
 }
 
 func TestPlanNotRetryableErr(t *testing.T) {
-	m := sync.Map{}
-	ctrl := gomock.NewController(t)
+	ctx, ctrl, test_context, distContext := testutil.InitTestContext(t, 2)
 	defer ctrl.Finish()
-	ctx := context.Background()
-	ctx = util.WithInternalSourceType(ctx, "dispatcher")
 
-	RegisterTaskMeta(t, ctrl, &m, getPlanNotRetryableErrDispatcherExt(ctrl))
-	distContext := testkit.NewDistExecutionContext(t, 2)
-	DispatchTaskAndCheckState(ctx, "key1", t, &m, proto.TaskStateFailed)
+	testutil.RegisterTaskMeta(t, ctrl, testutil.GetPlanNotRetryableErrDispatcherExt(ctrl), test_context)
+	testutil.DispatchTaskAndCheckState(t, ctx, "key1", test_context, proto.TaskStateFailed)
 	distContext.Close()
 }
