+++ pkg/disttask/framework/framework_pause_and_resume_test.go
@@ -16,7 +16,6 @@ package framework_test
 
 import (
 	"context"
-	"sync"
 	"testing"
 
 	"github.com/pingcap/failpoint"
@@ -24,10 +23,8 @@ import (
 	"github.com/pingcap/tidb/pkg/disttask/framework/handle"
 	"github.com/pingcap/tidb/pkg/disttask/framework/proto"
 	"github.com/pingcap/tidb/pkg/disttask/framework/storage"
-	"github.com/pingcap/tidb/pkg/testkit"
+	"github.com/pingcap/tidb/pkg/disttask/framework/testutil"
 	"github.com/stretchr/testify/require"
-	"github.com/tikv/client-go/v2/util"
-	"go.uber.org/mock/gomock"
 )
 
 func CheckSubtasksState(ctx context.Context, t *testing.T, taskID int64, state proto.TaskState, expectedCnt int64) {
@@ -42,23 +39,19 @@ func CheckSubtasksState(ctx context.Context, t *testing.T, taskID int64, state p
 }
 
 func TestFrameworkPauseAndResume(t *testing.T) {
-	var m sync.Map
-	ctrl := gomock.NewController(t)
+	ctx, ctrl, test_context, distContext := testutil.InitTestContext(t, 3)
 	defer ctrl.Finish()
-	ctx := context.Background()
-	ctx = util.WithInternalSourceType(ctx, "dispatcher")
 
-	RegisterTaskMeta(t, ctrl, &m, getMockBasicDispatcherExt(ctrl))
-	distContext := testkit.NewDistExecutionContext(t, 3)
+	testutil.RegisterTaskMeta(t, ctrl, testutil.GetMockBasicDispatcherExt(ctrl), test_context)
 	// 1. dispatch and pause one running task.
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/pauseTaskAfterRefreshTask", "2*return(true)"))
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/syncAfterResume", "return()"))
-	DispatchTaskAndCheckState(ctx, "key1", t, &m, proto.TaskStatePaused)
+	testutil.DispatchTaskAndCheckState(t, ctx, "key1", test_context, proto.TaskStatePaused)
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/pauseTaskAfterRefreshTask"))
 	// 4 subtask dispatched.
 	require.NoError(t, handle.ResumeTask(ctx, "key1"))
 	<-dispatcher.TestSyncChan
-	WaitTaskExit(ctx, t, "key1")
+	testutil.WaitTaskExit(t, ctx, "key1")
 	CheckSubtasksState(ctx, t, 1, proto.TaskStateSucceed, 4)
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/syncAfterResume"))
 
@@ -71,13 +64,13 @@ func TestFrameworkPauseAndResume(t *testing.T) {
 	// 2. pause pending task.
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/pausePendingTask", "2*return(true)"))
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/syncAfterResume", "1*return()"))
-	DispatchTaskAndCheckState(ctx, "key2", t, &m, proto.TaskStatePaused)
+	testutil.DispatchTaskAndCheckState(t, ctx, "key2", test_context, proto.TaskStatePaused)
 
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/pausePendingTask"))
 	// 4 subtask dispatched.
 	require.NoError(t, handle.ResumeTask(ctx, "key2"))
 	<-dispatcher.TestSyncChan
-	WaitTaskExit(ctx, t, "key2")
+	testutil.WaitTaskExit(t, ctx, "key2")
 	CheckSubtasksState(ctx, t, 1, proto.TaskStateSucceed, 4)
 	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/disttask/framework/dispatcher/syncAfterResume"))
