+++ pkg/disttask/framework/scheduler/BUILD.bazel
@@ -54,6 +54,7 @@ go_test(
         "//pkg/disttask/framework/mock/execute",
         "//pkg/disttask/framework/proto",
         "//pkg/disttask/framework/storage",
+        "//pkg/disttask/framework/testutil",
         "//pkg/kv",
         "//pkg/resourcemanager/pool/spool",
         "//pkg/resourcemanager/util",
@@ -61,6 +62,7 @@ go_test(
         "//pkg/util/logutil",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//util",
         "@org_golang_google_grpc//codes",
