+++ pkg/disttask/framework/scheduler/scheduler.go
@@ -231,12 +231,14 @@ func (s *BaseScheduler) run(ctx context.Context, task *proto.Task) (resErr error
 			continue
 		}
 		if subtask == nil {
+			failpoint.Inject("breakInSchedulerUT", func() {
+				failpoint.Break()
+			})
 			newTask, err := s.taskTable.GetGlobalTaskByID(runCtx, task.ID)
 			if err != nil {
 				logutil.Logger(s.logCtx).Warn("GetGlobalTaskByID meets error", zap.Error(err))
 				continue
 			}
-			// ywq todo how to break the loop.
 			// When the task move to next step or task state changes, the scheduler should exit.
 			if newTask.Step != task.Step || newTask.State != task.State {
 				break
