+++ pkg/disttask/framework/testutil/BUILD.bazel
@@ -2,7 +2,27 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "testutil",
-    srcs = ["util.go"],
+    srcs = [
+        "context.go",
+        "dispatcher_util.go",
+        "disttest_util.go",
+        "scheduler_util.go",
+    ],
     importpath = "github.com/pingcap/tidb/pkg/disttask/framework/testutil",
     visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/disttask/framework/dispatcher",
+        "//pkg/disttask/framework/dispatcher/mock",
+        "//pkg/disttask/framework/mock",
+        "//pkg/disttask/framework/mock/execute",
+        "//pkg/disttask/framework/proto",
+        "//pkg/disttask/framework/scheduler",
+        "//pkg/disttask/framework/storage",
+        "//pkg/domain/infosync",
+        "//pkg/testkit",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//util",
+        "@org_uber_go_mock//gomock",
+    ],
 )
